pyproject.toml
[project]
name = "grok2api"
version = "1.5.0"
description = "Grok2API rebuilt with FastAPI, fully aligned with the latest web call format. Supports streaming and non-streaming chat, image generation/editing, deep thinking, token pool concurrency, and automatic load balancing."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.3",
    "aiohttp-socks>=0.11.0",
    "aiomysql>=0.2.0",
    "asyncpg>=0.31.0",
    "curl-cffi>=0.13.0",
    "fastapi>=0.119.0",
    "greenlet>=3.3.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "livekit>=1.0.25",
    "loguru>=0.7.3",
    "pillow>=11.0.0",
    "orjson>=3.11.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.46",
    "tomli>=2.4.0",
    "uvicorn>=0.37.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]
